// RickLocalization/ClientApp/src/app/shared/models/card.model.ts
export class CardModel {
  personagemId1: number;
  personagemNome1: string;
  personagemDimensaoId1: number;
  personagemDimensao1: string;
  imagemPersonagem1: string;
  personagemId2: number;
  personagemNome2: string;
  personagemDimensaoId2: number;
  personagemDimensao2: string;
  imagemPersonagem2: string;
}